WebMIDIAPITest/webmidiapi_04_onmessage.html

<!DOCTYPE HTML>
<html>
  <head>
    <meta charset="UTF-8">
    <title>Web MIDI API Test: onmessage()</title>
    <link rel="author" title="Ryoya KAWAI" href="mailto:ryoya.kawai@gmail.com">
    <link rel="help" href="http://webaudio.github.io/web-midi-api/#obtaining-access-to-midi-devices">
    <link rel="match" href="reference/REFERENCE-FILE-NAME">
    <meta name="flags" content="WebMIDIAPI.js">
    <meta name="assert" content="Success when obtain over <span id='ulimit'></span> MIDI messages of noteOn or noteOff over onmidimessage() in 5 seconds.">
    <script src="resources/testharness.js"></script>
    <script src="resources/testharnessreport.js"></script>
    <script src="Uint8Array.js"></script>
    <script src="WebMIDIAPI.js"></script>

    <style type="text/css">
    /* ADD STYLE BLOCK HERE (PREFERRABLE TO INLINE STYLES) */
    </style>
  </head>
  <body>
    <p>Success when obtain over <span id="ulimit"></span> MIDI messages of noteOn or noteOff over onmidimessage() in 5 seconds.</p>
    <div id="selectdevice"></div>
    <p id="msgstatus">Now <span id="msgcnt">0 message</span> accepted.</p>
    <div id="accept" style="color:red;"></div>
    <div id="warning" style="color:red;"></div>
    <div id="warning2" style="color:red;"></div>

    <div id="log"></div>

    <!-- PAGE CONTENT -->
    <script type="text/javascript">
    var MIDI;
    var timerId={ };
    var getPort={ };
    var uLimit=8;
    var inputCount=0;
    var acceptInput=["^[0-9]+\\.KEYBOARD$", "^[0-9]+\\.USB ?", "^[0-9]+\\.nanoKEY2$", "^[0-9]+\\.?"];
    var accept=null;

    document.getElementById("ulimit").innerHTML=uLimit;
    document.getElementById("msgcnt").innerHTML=inputCount + "/" + uLimit +  " messages";

    var test_inputs = async_test(
        "inputs() - W3C Spec: Web MIDI API",
        {timeout: 5000} //msec
    );
    navigator.requestMIDIAccess().then( successCallback, errorCallback );
    function successCallback(access) {
        MIDI = access;
        test_inputs.step(function() {
            var inputs=MIDI.inputs();
            console.log("[InputDevices] ", inputs);
            if(inputs.length==0) {
                document.getElementById("warning").innerHTML="Your Computer looks no <b><u>MIDI INPUT</u></b> devices are pluged-in.<br> Please plug-in at least one <b><u>MIDI INPUT</u></b> device before run this Test.";
                assert_false(true, "[No MIDI input devices are connected.]");
                test_inputs.done();
            } else {
                var input = document.createElement("input");
                input.type="button"; input.id="select"; input.value="Select & Run";
                var select = document.createElement("select");
                select.id="midiin";
                for(var i=0; i<inputs.length; i++) {
                    select.options[i]=new Option(inputs[i]["name"]+" ("+inputs[i]["id"]+")", i);
                }
                document.getElementById("selectdevice").appendChild(select);
                document.getElementById("selectdevice").appendChild(input);
            }

            document.getElementById("select").addEventListener("click", function() {
                disableButton()
                var selIdx=document.getElementById("midiin").selectedIndex;
                var i = inputs[selIdx];
                assert_true(inputs instanceof Object);
                console.log(i, selIdx);

                i.onmidimessage=i.onmessage=function(event) {
                    if(inputCount<uLimit) console.log(event.data[0].toString(16) + " " + event.data[1].toString(16) + " " + event.data[2].toString(16));
                    document.getElementById("msgcnt").innerHTML=inputCount + "/" + uLimit +  " messages";
                    if(event.data[0].toString(16).substr(0,1)==9 || event.data[0].toString(16).substr(0,1)==8) inputCount++;
                    if(inputCount>uLimit) {
                        document.getElementById("msgstatus").style.setProperty("display", "none");
                        disableButton();
                        test_inputs.done();
                    }
                };

            });

        });
    }
    function errorCallback(message) {
        test_inputs.step(function() {
            assert_true(false);
            test_inputs.done();
        });
    }
    function disableButton() {
        document.getElementById("midiin").setAttribute("disabled", "disabled");
        document.getElementById("select").setAttribute("disabled", "disabled");
    }


    </script>

  </body>
</html>
